Extras/chatter-rag-app/backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asgiref==3.8.1
    # via opentelemetry-instrumentation-asgi
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.2.0
    # via aiohttp
backoff==2.2.1
    # via posthog
bcrypt==4.1.2
    # via chromadb
blinker==1.7.0
    # via flask
build==1.2.1
    # via chromadb
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   pulsar-client
    #   requests
charset-normalizer==3.3.2
    # via requests
chroma-hnswlib==0.7.3
    # via chromadb
chromadb==0.4.24
    # via langchain-chroma
click==8.1.7
    # via
    #   flask
    #   typer
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
exceptiongroup==1.2.0
    # via anyio
fastapi==0.110.1
    # via
    #   chromadb
    #   langchain-chroma
filelock==3.13.4
    # via huggingface-hub
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==4.0.1
    # via -r requirements.in
flatbuffers==24.3.25
    # via onnxruntime
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.3.1
    # via huggingface-hub
google-auth==2.29.0
    # via kubernetes
googleapis-common-protos==1.63.0
    # via opentelemetry-exporter-otlp-proto-grpc
greenlet==3.0.3
    # via sqlalchemy
grpcio==1.62.1
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via ollama
huggingface-hub==0.20.3
    # via tokenizers
humanfriendly==10.0
    # via coloredlogs
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==7.0.0
    # via opentelemetry-api
importlib-resources==6.4.0
    # via chromadb
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via flask
jsonpatch==1.33
    # via langchain-core
jsonpointer==2.4
    # via jsonpatch
kubernetes==29.0.0
    # via chromadb
langchain==0.1.19
    # via -r requirements.in
langchain-chroma==0.1.0
    # via -r requirements.in
langchain-community==0.0.38
    # via
    #   -r requirements.in
    #   langchain
langchain-core==0.1.52
    # via
    #   langchain
    #   langchain-chroma
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.0.1
    # via langchain
langsmith==0.1.48
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.21.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
mmh3==4.1.0
    # via chromadb
monotonic==1.6
    # via posthog
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.24.4
    # via
    #   chroma-hnswlib
    #   chromadb
    #   langchain
    #   langchain-chroma
    #   langchain-community
    #   onnxruntime
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
ollama==0.1.9
    # via -r requirements.in
onnxruntime==1.17.3
    # via chromadb
opentelemetry-api==1.24.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-exporter-otlp-proto-common==1.24.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.24.0
    # via chromadb
opentelemetry-instrumentation==0.45b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi==0.45b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.45b0
    # via chromadb
opentelemetry-proto==1.24.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.24.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.45b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-util-http==0.45b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
orjson==3.10.1
    # via
    #   chromadb
    #   langsmith
overrides==7.7.0
    # via chromadb
packaging==23.2
    # via
    #   build
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   onnxruntime
posthog==3.5.0
    # via chromadb
protobuf==4.25.3
    # via
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
pulsar-client==3.5.0
    # via chromadb
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.7.0
    # via
    #   chromadb
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.18.1
    # via pydantic
pygments==2.17.2
    # via rich
pypika==0.48.9
    # via chromadb
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.9.0.post0
    # via
    #   kubernetes
    #   posthog
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.1
    # via
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
requests==2.31.0
    # via
    #   -r requirements.in
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   langchain
    #   langchain-community
    #   langsmith
    #   posthog
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rich==13.7.1
    # via typer
rsa==4.9
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   kubernetes
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.29
    # via
    #   langchain
    #   langchain-community
starlette==0.37.2
    # via fastapi
sympy==1.12
    # via onnxruntime
tenacity==8.2.3
    # via
    #   chromadb
    #   langchain
    #   langchain-community
    #   langchain-core
tokenizers==0.15.2
    # via chromadb
tomli==2.0.1
    # via
    #   build
    #   pyproject-hooks
tqdm==4.66.4
    # via
    #   -r requirements.in
    #   chromadb
    #   huggingface-hub
typer==0.12.3
    # via chromadb
typing-extensions==4.11.0
    # via
    #   anyio
    #   asgiref
    #   chromadb
    #   fastapi
    #   huggingface-hub
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.1
    # via
    #   kubernetes
    #   requests
uvicorn[standard]==0.29.0
    # via chromadb
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websocket-client==1.7.0
    # via kubernetes
websockets==12.0
    # via uvicorn
werkzeug==3.0.2
    # via flask
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
yarl==1.9.4
    # via aiohttp
zipp==3.18.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
